+++ system/core/Input.php
@@ -383,27 +383,7 @@ class CI_Input {
 	 */
 	public function valid_ip($ip)
 	{
-		if ($this->ip_version($ip) === 4)
-		{
-			return (bool) filter_var($ip, FILTER_VALIDATE_IP, FILTER_FLAG_IPV4);
-		}
-		else
-		{
-			return (bool) filter_var($ip, FILTER_VALIDATE_IP, FILTER_FLAG_IPV6);
-		}
-	}
-
-	// --------------------------------------------------------------------
-
-	/**
-	 * Return ip version
-	 *
-	 * @param	string
-	 * @return	int
-	 */
-	public function ip_version($ip)
-	{
-		return strpos($ip, ":") === FALSE ? 4 : 6;
+		return (bool) filter_var($ip, FILTER_VALIDATE_IP);
 	}
 
 	// --------------------------------------------------------------------